components/api-backend/pyproject.toml
[project]
name = "api-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi[all]>=0.115.13",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
